+++ Sources/MapboxCoreNavigation/CoreNavigationNavigator.swift
@@ -147,9 +147,9 @@ class Navigator {
 
 extension Navigator: ElectronicHorizonObserver {
     public func onPositionUpdated(for position: ElectronicHorizonPosition, distances: [String : MapboxNavigationNative.RoadObjectDistanceInfo]) {
-        let userInfo: [ElectronicHorizon.NotificationUserInfoKey: Any] = [
+        let userInfo: [RoadGraph.NotificationUserInfoKey: Any] = [
             .positionKey: RoadGraph.Position(position.position()),
-            .treeKey: ElectronicHorizon(position.tree()),
+            .treeKey: try! position.tree(),
             .updatesMostProbablePathKey: position.type() == .UPDATE,
             .distancesByRoadObjectKey: distances.mapValues(RoadObjectDistanceInfo.init),
         ]
@@ -157,7 +157,7 @@ extension Navigator: ElectronicHorizonObserver {
     }
     
     public func onRoadObjectEnter(for info: RoadObjectEnterExitInfo) {
-        let userInfo: [ElectronicHorizon.NotificationUserInfoKey: Any] = [
+        let userInfo: [RoadGraph.NotificationUserInfoKey: Any] = [
             .roadObjectIdentifierKey: info.roadObjectId,
             .didTransitionAtEndpointKey: info.isEnterFromStartOrExitFromEnd,
         ]
@@ -165,7 +165,7 @@ extension Navigator: ElectronicHorizonObserver {
     }
     
     public func onRoadObjectExit(for info: RoadObjectEnterExitInfo) {
-        let userInfo: [ElectronicHorizon.NotificationUserInfoKey: Any] = [
+        let userInfo: [RoadGraph.NotificationUserInfoKey: Any] = [
             .roadObjectIdentifierKey: info.roadObjectId,
             .didTransitionAtEndpointKey: info.isEnterFromStartOrExitFromEnd,
         ]