+++ Sources/MapboxCoreNavigation/EHorizon/RoadGraph.swift
@@ -3,7 +3,7 @@ import Turf
 import MapboxNavigationNative
 
 /**
- `RoadGraph` provides methods to get edge shape (e.g. `ElectronicHorizon.Edge`) and metadata.
+ `RoadGraph` provides methods to get edge shape (e.g. `RoadGraph.Edge`) and metadata.
  
  You do not create a `RoadGraph` object manually. Instead, observe the `Notification.Name.electronicHorizonDidUpdatePosition` notification to obtain edge identifiers and get more details about the edges using the `RouteController.roadGraph` or `PassiveLocationDataSource.roadGraph` property.
  */
@@ -14,9 +14,9 @@ public final class RoadGraph {
      
      - returns: Metadata about the edge with the given edge identifier, or `nil` if the edge is inaccessible.
      */
-    public func edgeMetadata(edgeIdentifier: ElectronicHorizon.Edge.Identifier) -> ElectronicHorizon.Edge.Metadata? {
+    public func edgeMetadata(edgeIdentifier: Edge.Identifier) -> Edge.Metadata? {
         if let edgeMetadata = native.getEdgeMetadata(forEdgeId: UInt64(edgeIdentifier)) {
-            return ElectronicHorizon.Edge.Metadata(edgeMetadata)
+            return Edge.Metadata(edgeMetadata)
         }
         return nil
     }
@@ -26,7 +26,7 @@ public final class RoadGraph {
      
      - returns: A line string corresponding to the given edge identifier, or `nil` if the edge is inaccessible.
      */
-    public func edgeShape(edgeIdentifier: ElectronicHorizon.Edge.Identifier) -> LineString? {
+    public func edgeShape(edgeIdentifier: Edge.Identifier) -> LineString? {
         guard let locations = native.getEdgeShape(forEdgeId: UInt64(edgeIdentifier)) else {
             return nil
         }